requirements.txt
appnope==0.1.3
asttokens==2.0.7
backcall==0.2.0
black==22.6.0
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_83242e7e-f82d-4a71-8ef2-9d71d212d249gu_wxmeq/croots/recipe/certifi_1655968827803/work/certifi
cfgv==3.3.1
click==8.1.3
decorator==5.1.1
distlib==0.3.5
executing==0.9.1
filelock==3.7.1
identify==2.5.3
ipdb==0.13.9
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
nodeenv==1.7.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
pre-commit==2.20.0
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
pycryptodome==3.15.0
Pygments==2.12.0
PyYAML==6.0
six==1.16.0
stack-data==0.3.0
toml==0.10.2
tomli==2.0.1
traitlets==5.3.0
virtualenv==20.16.3
wcwidth==0.2.5
